requirements.txt
asgiref==3.8.1
asttokens==3.0.0
certifi==2024.12.14
charset-normalizer==3.4.0
colorama==0.4.6
decorator==5.1.1
Django==4.2
django-extensions==3.2.3
django-seed==0.3.1
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
executing==2.1.0
Faker==33.1.0
idna==3.10
ipython==8.30.0
jedi==0.19.2
matplotlib-inline==0.1.7
parso==0.8.4
pillow==11.0.0
prompt_toolkit==3.0.48
psycopg2==2.9.10
pure_eval==0.2.3
Pygments==2.18.0
python-dateutil==2.9.0.post0
requests==2.32.3
setuptools==75.1.0
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
toposort==1.10
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
wcwidth==0.2.13
wheel==0.44.0
